Fix #1791 A user will still be created if the validation e-mail cannot be sent
